Support filtering files and sorting output by iOS domain
Some relative paths are not globally unique, so in order to match
them the domain of the application that created them can be used to
restrict the matching of files.
This also adds the ability to sort extracted files into subfolders
by their domain, to avoid clashes from non unique relative paths.

In theory, this means that it is now possible to extract all files
from a backup using one command:

    backup.extract_files(relative_paths_like="%", domain_like="%",
                         output_folder="./output",
                         preserve_folders=True,
                         domain_subfolders=True)

I have not tested this, use at own risk ¯\_(ツ)_/¯