Add a Dockerfile for a pre-built container version
This was more about an excuse to play with the GitHub Container
Registry than of immediate practical benefit; but it may help on
Windows where fastpbkdf2 is a pain.
It may also reassure anyone worried about the integrity of their
backup since, as in the example, the backup can be mounted as
read-only and the output folder treated separately.