Decrypt files in chunks to avoid memory exhaustion with large files
This only works when decrypting files to disk, and so a lot of
refactoring was necessary to separate the two code paths.
This now feels cleaner, and there is more code re-use via a new
utility class for PLists and some splitting of the existing internal
functions. The public API has not changed.

There's a beautiful walrus operator to do the looping through byte
chunks of the files; good job Python 3.8 was already required! This
is safe because only empty byte strings are false-y; all other
values including zeros and null bytes keep the loop going.